requirements.txt
# Requirements for Label Maker Application - PythonAnywhere Optimized
# This file contains all dependencies needed to run the application on PythonAnywhere

# Core Flask framework and extensions
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3

# HTTP requests for API calls
requests==2.31.0

# Data processing and analysis
pandas==2.0.3
numpy==1.24.3
python-dateutil==2.8.2

# Document processing
python-docx==0.8.11
docxtpl==0.16.7
docxcompose==1.4.0
openpyxl==3.1.2

# PDF generation
reportlab==4.0.4

# Image processing
Pillow==9.5.0

# File monitoring and system utilities
watchdog==3.0.0

# Additional utilities
six==1.16.0
pytz==2023.3

flask-caching


